chore(demos): Build out history demos
Moved demos from header section and added some more detail.
