+++ packages/module/src/ChatbotConversationHistoryNav/ChatbotConversationHistoryNav.scss
@@ -63,6 +63,7 @@
     --pf-v6-c-drawer__panel--PaddingBlockEnd: var(--pf-t--global--spacer--lg);
     --pf-v6-c-drawer__panel--RowGap: var(--pf-t--global--spacer--lg);
     overflow-x: hidden;
+    overflow-y: hidden;
   }
 
   // Drawer head
@@ -145,10 +146,13 @@
 // ============================================================================
 .pf-chatbot--embedded {
   .pf-chatbot__history.pf-v6-c-drawer {
-    height: 100%;
+    height: inherit;
     .pf-v6-c-drawer__body {
       width: 100%;
-      min-height: 100%;
+      height: 100%;
+    }
+    .pf-v6-c-drawer__panel {
+      height: 100%;
     }
   }
 }
